+++ src/index.js
@@ -29,4 +29,4 @@ export { default as Filter } from './Filter'
 export { default as Modal } from './Modal'
 export { default as Message } from './Message'
 
-export { default as DataHolder } from './DataHolder'
+export { default as ValuesHolder } from './ValuesHolder'