+++ src/utils/strings.js
@@ -17,12 +17,16 @@ export function format () {
 
 export function substitute (str, obj) {
   if (typeof str === 'string') {
-    return str.replace((/\\?\{([^{}]+)\}/g), function (match, name) {
-      if (match.charAt(0) === '\\') {
-        return match.slice(1)
-      }
-      return (obj[name] === null || obj[name] === undefined) ? '' : obj[name]
-    })
+    if (str.indexOf('{') < 0) {
+      return obj[str]
+    } else {
+      return str.replace((/\\?\{([^{}]+)\}/g), function (match, name) {
+        if (match.charAt(0) === '\\') {
+          return match.slice(1)
+        }
+        return (obj[name] === null || obj[name] === undefined) ? '' : obj[name]
+      })
+    }
   } else if (typeof str === 'function') {
     return str(obj)
   }