+++ CHANGELOG.md
@@ -6,8 +6,9 @@
  - Select移除数据处理，交由FormItem处理，增加lazy render
  - Message拆分，去除PubSub依赖，暴露success, warning, error接口，UI调整到顶部
  - Modal拆分，去除PubSub依赖，改为垂直居中
- - Tree折线改为SVG
+ - Tree折线改为SVG，greedy改为capture
  - 增加Alert
+ - Table 拆分，Sort，Pagination改为高阶组件，去除state
 
 # v0.6.1
  - dataSource 变为了 [fetch](http://lobos.github.io/react-ui/#/fetch)。方式改为higher order component，使用[refetch](https://github.com/Lobos/react-ui)，支持ajax，jsonp，cache